rewrite item parsing, part 1/2
This parses items from the wonderful EldenRingDatabase/erdb, which
parses them automatically from the games files. This is a lot more
reliable than me going through the savefile and attempting to find IDs
manually. This also makes many more items be supported than before.

With this new item detection the current model of parsing does not make
much sense anymore, and will need to be redone in part two of this
rewrite.